+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added support finding R executables in VSCode using `Rpath` settings in the R extension. (#3378)
+- The Publisher logs view automatically opens when a deployment starts. (#3235)
+- The Publisher logs automatically scroll to follow new log entries during deployment. (#2997)
+- Publisher Log stages automatically expand when running and collapse when completed. (#2998)
 - Added R, Python, and Quarto interpreter version to the logs.
 
 ### Fixed